python/rec_service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi-cli==0.0.3
filelock==3.14.0
Flask==3.0.3
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keybert==0.8.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging==24.0
pillow==10.3.0
praw==7.7.1
prawcore==2.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.4
redisvl==0.2.0
regex==2024.5.15
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
update-checker==0.18.0
urllib3==2.1.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.3
